+++ crates/nostr-relay-pool/src/relay/error.rs
@@ -60,6 +60,8 @@ pub enum Error {
     NotReady,
     /// Relay not connected
     NotConnected,
+    /// Received termination request
+    TerminationRequest,
     /// Received shutdown
     ReceivedShutdown,
     /// Relay message
@@ -146,6 +148,7 @@ impl fmt::Display for Error {
             }
             Self::NotReady => write!(f, "relay is initialized but not ready"),
             Self::NotConnected => write!(f, "relay not connected"),
+            Self::TerminationRequest => write!(f, "received termination request"),
             Self::ReceivedShutdown => write!(f, "received shutdown"),
             Self::RelayMessage(message) => write!(f, "{message}"),
             Self::BatchMessagesEmpty => write!(f, "can't batch empty list of messages"),